+++ components/JsonLdLandingFaq.tsx
@@ -1,10 +1,11 @@
 import { buildLandingFaqPageJsonLd } from '@/lib/landing-faq';
+import { jsonLdString } from '@/lib/json-ld';
 
 export function JsonLdLandingFaq() {
   return (
     <script
       type="application/ld+json"
-      dangerouslySetInnerHTML={{ __html: JSON.stringify(buildLandingFaqPageJsonLd()) }}
+      dangerouslySetInnerHTML={{ __html: jsonLdString(buildLandingFaqPageJsonLd()) }}
     />
   );
 }